requirements.txt
pillow
pdfminer.six
python-docx